# perf: Moved  `Authorization Control` Controller Logic To Utilities