--- a/erpnext/setup/doctype/authorization_control/README.md
+++ b/erpnext/setup/doctype/authorization_control/README.md
@@ -1,1 +1,0 @@
-Tool to define Authorization Rules (limits above which only specified roles have right to Submit).