… erpnext/setup/doctype/authorization_control/authorization_control.json
… erpnext/setup/doctype/authorization_control/authorization_control.py