@sagarvora 2024-03-28
#### Performance Improvements
- No extra DB call to get doc values during transaction submission
- No overhead of DocType meta
- No need to initialize